bot iteration with middleware

+++ internal/models/user.go
@@ -5,7 +5,7 @@ import "gorm.io/gorm"
 type User struct {
 	gorm.Model
 	Username   string `gorm:"unique;not null"`
-	TelegramId int    `gorm:"unique;not null"`
+	TelegramId int64  `gorm:"unique;not null"`
 
 	Wishes []Wish
 }
